Decided on a new advanaced processing approach, deleting old one but moving some functions to data_utils.py
